[GREEN] - Pasa la prueba de terminar la asociación entre una alarma del tipo B y un autor, cuando las condiciones así lo ameritan.

+++ Obligatorio1_SantiagoAlpuy/BusinessLogic/Controllers/AlertBAuthorController.cs
@@ -38,13 +38,19 @@ namespace BusinessLogic.Controllers
         public void RemoveAssociationAlertAuthor(AlertB alert, Author author)
         {
             AlertBAuthor alertAuthor = FindAssociationAlertAuthor(alert, author);
-            repository.Remove(alertAuthor);
+            if (alertAuthor != null)
+                repository.Remove(alertAuthor);
         }
 
         public ICollection<AlertBAuthor> GetAllRelationsByAlertId(int alertBId)
         {
             ICollection<AlertBAuthor> collection = repository.GetEntitiesByPredicate(x => x.AlertBId == alertBId);
             return collection;
+        }
+
+        public void RemoveAll()
+        {
+            repository.ClearAll();
         }
 
     }
